Added baseline as an option for drivers to pick.  Lag cause found.
Lag issue on second robot found!  In the robot.java teleopPeriodic()
function, commenting out elevator.elevatorPeriodic() and
intake.intakePeriodic() fixed it.  I assume this is because they both
try to read from encoders that aren't plugged in.  The issue was also
happening on the first bot, and I suspect (but can't prove until Utica)
that it was because the encoder was plugged into the wrong talon
(elevator2 vs elevator1).
Put baseline in scale auto because we don't have code for the scale.
Confirmed that reading current draw from PDP works.