Utica week 1 competition
Auto: The first robot can drive straight, if all the talons are
assembled correctly.  Added a delay for baseline execution where it is
read from slider 0 and the input can only be 0-10.  Only baseline was
used the entire competition and this is an issue.
Elevator: Put brake on automatic then quickly changed it to only when
the X button was held because it may have broken the chain instead of
braking it.
The limit switch was never physically added so the encoder reset was
being called on every loop of the program, causing the annoying lag.
Drive: Put brake mode back on and multiplied controller inputs by .8 so
the max speed is 80% because running into things at high speeds without
stopping is bad.
General: I experienced 2 instances on the competition field where the
driver station would take a really long time to connect to the FMS and
robot.  It was fixed both times by rebooting the computer or turning on
and using the spare.  In the future, it should be rebooted/replaced
before the day starts, if there is a break, and before
eliminations/finals.