Many autonomous improvements.
Lots of functions to abstract away unnecessary details.  It all comes
together in the new drive function.  This will make new auto development
far easier.
Made the mirror variable a global variable that lives in the autonomous
class so it can be easily read there, but is changed in robot.java where
the autonomous function selection happens.
Improved names and parameters of intake and elevator functions.
Added a test mode in auto for development of crazy things. Currently it
is testing the new drive function and raising the elevator while first
driving out (change that number).
If anyone is curious, the wrist motor was briefly changed to a piston,
which would have made things easier.  Then it was changed back to a
motor.
Plan for center switch: start to the right of the exchange (it is offset
left of center for red and blue) with the robot directly in front of the
right switch.  If that is our side, we can just drive forward and
deposit in it.  If ours is on the left, we will have to drive to the
left a bit.  Once that is down, that cross code can be adapted for when
we start on the side - we could cross the field in front of the switch,
maybe with a delay, to deposit in the switch if it is on the side we
don't start on.  This will mean we don't have to worry about mirroring,
instead we will have to write 2 different functions.