Autonomous is now ready to write 2 cube autos.
Delay for all autos is working.  Baseline is finally in the switch
statement format.  Moved the set power to motors line to the end of
autonomous periodic in robot.java so it won't have to be in every
autonomous function.  Fixed a bug where state was initialized but not
set to a value (did you know that integers default to zero?) in the
process of making the delay work.

Congratulations Michael, you get 90% drive power now.

Got rid of the elevator limit switch because it never existed in the
first place.