Center second cube auto with right angle turns.
Will be changed to about 45 degree turn to save time and movements.
Center two cubes is no longer a default.
Drive encoder average distance is displayed when disabled.
Motors are put in coast mode when disabled.  For competition, this
should be deleted so we don't fall off a ramp.  Instead, it will be in
coast mode only at robot init so that it can be pushed onto the field
with ease.